requirements.txt
Flask
python-dotenv
langchain
langchain-openai
langchain-community
chromadb
PyPDF2
sentence-transformers